LayoutTests/accessibility/canvas-fallback-content-2-expected.txt

Link  Button  Button  Button
Focusable
Heading

ARIA button
ARIA disabled button
ARIA enabled button
ARIA required button
ARIA toggle button
ARIA link
This tests a number of different elements in canvas fallback content to make sure their accessible attributes are essentially identical to the corresponding elements outside of canvas fallback content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


link1
link2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

button1
button2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

labeled-button1
labeled-button2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

button-with-title1
button-with-title2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

text1
text2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

checkbox1
checkbox2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

number1
number2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

radio1
radio2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

slider1
slider2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

submit1
submit2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

combobox1
combobox2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

listbox1
listbox2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

textarea1
textarea2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

focusable1
focusable2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

heading1
heading2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

aria-button1
aria-button2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

aria-disabledbutton1
aria-disabledbutton2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

aria-enabledbutton1
aria-enabledbutton2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

aria-requiredbutton1
aria-requiredbutton2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

aria-togglebutton1
aria-togglebutton2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

aria-link1
aria-link2
PASS axElement2.role is axElement1.role
PASS axElement2.roleDescription is axElement1.roleDescription
PASS axElement2.title is axElement1.title
PASS axElement2.description is axElement1.description
PASS axElement2.helpText is axElement1.helpText
PASS axElement2.stringValue is axElement1.stringValue
PASS axElement2.isEnabled is axElement1.isEnabled
PASS axElement2.isRequired is axElement1.isRequired
PASS axElement2.isChecked is axElement1.isChecked
PASS axElement2.intValue is axElement1.intValue
PASS axElement2.minValue is axElement1.minValue
PASS axElement2.maxValue is axElement1.maxValue

PASS successfullyParsed is true

TEST COMPLETE
